refactor: move getApiAccessToken function to auth API and remove old utility file
